fix: centralize test type guards
